Added missing cast.

+++ src/Simulation.cpp
@@ -139,7 +139,7 @@ std::vector<int> Simulation::getBitsDistribution(
 int Simulation::getCurrentDeathChanceInPercent(int populationCount) const
 {
     const float populationRatio{static_cast<float>(populationCount) /
-                                params_.population_.max_};
+                                static_cast<float>(params_.population_.max_)};
     const float deathChance{populationRatio * config::maxPercentage};
     return static_cast<int>(std::round(deathChance));
 }
